refactor: Employee 목록 커서 생성 책임 이동
